Move modals to separate folder
