+++ src/js/components/stock-movement-wizard/modals/SubstitutionsModal.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ModalWrapper from '../form-elements/ModalWrapper';
+import ModalWrapper from '../../form-elements/ModalWrapper';
 
 const SubstitutionsModal = () => (
   <ModalWrapper title="Substitutions" >
